+++ robotframework-ls/docs/changelog.md
@@ -1,3 +1,9 @@
+NEXT
+-----------------------------
+
+- New setting `"robot.completions.keywords.argumentsSeparator"` allows customizing spacing betwen keyword and arguments when auto-completing Keyword with arguments. [#683](https://github.com/robocorp/robotframework-lsp/issues/683)
+
+
 New in 0.48.0 (2022-05-30)
 -----------------------------